[#121] feat: Bookmark에 unique 지정, Document 연관 관계, 정적 팩토리 메서드, member 동일 확인 메서드 추가 - Member, Document 복합 unique 키 설정 - Member 삭제 시 Bookmark 삭제되도록 CascadeType.REMOVE, orpahnRemoval 설정 - Document는 내용만 삭제 예정으로 삭제 시는 따로 구현하지 않음.
