+++ src/main/java/goorm/eagle7/stelligence/domain/bookmark/model/Bookmark.java
@@ -1,0 +1,52 @@
+package goorm.eagle7.stelligence.domain.bookmark.model;
+
+import goorm.eagle7.stelligence.common.entity.BaseTimeEntity;
+import goorm.eagle7.stelligence.domain.document.content.model.Document;
+import goorm.eagle7.stelligence.domain.member.model.Member;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import jakarta.persistence.UniqueConstraint;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(uniqueConstraints = {
+	@UniqueConstraint(name = "UniqueMemberAndDocument", columnNames = {"member_id", "document_id"})})
+public class Bookmark extends BaseTimeEntity {
+
+	@Id
+	@Column(name = "bookmark_id")
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "member_id")
+	private Member member;
+
+	// Document가 삭제될 때, 북마크 삭제 구현 X
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "document_id")
+	private Document document;
+
+	public static Bookmark of(Member member, Document document) {
+		Bookmark bookmark = new Bookmark();
+		bookmark.member = member;
+		bookmark.document = document;
+		return bookmark;
+	}
+
+	public boolean isSameMember(Long memberId) {
+		return this.member.getId().equals(memberId);
+	}
+
+}
