--- src/main/java/goorm/eagle7/stelligence/domain/member/model/Bookmark.java
@@ -1,34 +1,0 @@
-package goorm.eagle7.stelligence.domain.member.model;
-
-import goorm.eagle7.stelligence.common.entity.BaseTimeEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Bookmark extends BaseTimeEntity {
-
-	@Id
-	@Column(name = "bookmark_id")
-	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
-
-	@ManyToOne(fetch = FetchType.LAZY)
-	@JoinColumn(name = "member_id")
-	private Member member;
-
-	// @ManyToOne(fetch = FetchType.LAZY)
-	// @JoinColumn(name = "document_id")
-	// private Document document;
-
-}
